+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 * scolv
 
-  * ```locator.minimumDepth``` is now deprecated in favour if ```olv.locator.minimumDepth```
+  * ```locator.minimumDepth``` is now deprecated in favour of ```olv.locator.minimumDepth```
+  * ```olv.locator``` is now deprecated in favour of ```olv.locator.interface```
 
 
 ## Release 2017.124